+++ packages/node/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `block-confirmations` arg being removed in recent sync with main SDK (#213)
 
 ## [3.3.4] - 2023-11-16
 ### Fixed
@@ -12,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.3.3] - 2023-11-13
 ### Changed
-- Updates to match changes in 
+- Updates to match changes in `@subql/node-core` (#210)
   - Dictionary service to use dictionary registry
   - Use yargs from node core